+++ src/freenet/support/io/PersistentBlobTempBucketFactory.java
@@ -440,7 +440,7 @@ outer:		while(true) {
 			// FIXME limits size to Integer.MAX_VALUE entries so that we don't need to worry about freeBlocksCache.
 			// FIXME long term, make freeBlocksCache take a long.
 			if(blocks + addBlocks > Integer.MAX_VALUE) {
-				addBlocks = (blocks + addBlocks) - (long)Integer.MAX_VALUE;
+				addBlocks = (blocks + addBlocks) - Integer.MAX_VALUE;
 				if(addBlocks <= 0) return changedTags;
 			}
 			
@@ -732,7 +732,7 @@ outer:				while(true) {
 						PersistentBlobTempBucketTag tag = new PersistentBlobTempBucketTag(PersistentBlobTempBucketFactory.this, last);
 						container.store(tag);
 						synchronized(this) {
-							freeBlocksCache.setBit((int)last, false);
+							freeBlocksCache.setBit(last, false);
 						}
 						continue;
 					}
